[refactor] 네이밍 변경 (dto)
domain 패키지를 논리에 맞는 이름인 dto 로 변경
