+++ src/main/java/com/suitecare/suitecare/api/family/service/FamilyService.java
@@ -1,9 +1,9 @@
 package com.suitecare.suitecare.api.family.service;
 
 import com.suitecare.suitecare.api.family.repository.FamilyMapper;
-import com.suitecare.suitecare.api.family.domain.CreateFamilyDTO;
-import com.suitecare.suitecare.api.family.domain.FamilyDTO;
-import com.suitecare.suitecare.api.family.domain.LoginDTO;
+import com.suitecare.suitecare.api.family.dto.CreateFamilyDTO;
+import com.suitecare.suitecare.api.family.dto.FamilyDTO;
+import com.suitecare.suitecare.api.family.dto.LoginDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
